+++ tracy/inc/tpsa_lin.h
@@ -4,32 +4,33 @@
 
 void daeps_(const double eps);
 
-void danot_(const int no);
+void danot_(const long int no);
 
-void daini_(int no, int nv, int fio);
+void daini_(long int no, long int nv, long int fio);
 
-void lieini(const int no, const int nv, const int nd2i);
+void lieini(const long int no, const long int nv, const long int nd2i);
 
-void daall_(tps_buf &x, const int nd2, const char *daname,
-	    const int no, const int nv);
+void daall_(tps_buf &x, const long int nd2, const char *daname,
+	    const long int no, const long int nv);
 
-void dadal_(tps_buf &x, const int ss_dim);
+void dadal_(tps_buf &x, const long int ss_dim);
 
-void davar_(tps_buf &x, const double r, const int i);
+void davar_(tps_buf &x, const double r, const long int i);
 
 void dacon_(tps_buf &x, const double r);
 
-void dapek_(const tps_buf &x, const int jj[], double &r);
+void dapek_(const tps_buf &x, const long int jj[], double &r);
 
-void dapok_(tps_buf &x, const int jj[], const double r);
+void dapok_(tps_buf &x, const long int jj[], const double r);
 
-double getmat(const ss_vect<tps> &map, const int i, const int j);
+double getmat(const ss_vect<tps> &map, const long int i, const long int j);
 
-void putmat(ss_vect<tps> &map, const int i, const int j, const double r);
+void putmat(ss_vect<tps> &map, const long int i, const long int j,
+	    const double r);
 
-void getlinmat(const int nv, const ss_vect<tps> &map, Matrix &mat);
+void getlinmat(const long int nv, const ss_vect<tps> &map, Matrix &mat);
 
-void putlinmat(const int nv, const Matrix &mat, ss_vect<tps> &map);
+void putlinmat(const long int nv, const Matrix &mat, ss_vect<tps> &map);
 
 void dacop_(const tps_buf &x, tps_buf &z);
 
@@ -84,12 +85,14 @@ void daarctan(const tps_buf &x, tps_buf &z);
 
 void dafun_(const char *fun, const tps_buf &x, tps_buf &z);
 
-void dacct_(const ss_vect<tps> &x, const int i,
-	    const ss_vect<tps> &y, const int j, ss_vect<tps> &z, const int k);
+void dacct_(const ss_vect<tps> &x, const long int i,
+	    const ss_vect<tps> &y, const long int j, ss_vect<tps> &z,
+	    const long int k);
 
-void dainv_(const ss_vect<tps> &x, const int i, ss_vect<tps> &z, const int k);
+void dainv_(const ss_vect<tps> &x, const long int i, ss_vect<tps> &z,
+	    const long int k);
 
-void Rotmap(const int n, ss_vect<tps> &map, const Matrix &R);
+void Rotmap(const long int n, ss_vect<tps> &map, const Matrix &R);
 
 void daabs_(const tps_buf &x, double &r);
